% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mod.R
\docType{data}
\name{example_mod}
\alias{example_mod}
\title{Example Model Output}
\format{
A data.frame with the following variables:
\describe{
\item{\code{date}}{date as 'character' column class.}
\item{\code{streamflow_cfs}}{modeled streamflow in units of feet^3/second.}
\item{\code{Date}}{date as 'Date' column class.}
}
}
\usage{
example_mod
}
\description{
An example dataset with daily modeled (simulated) streamflow.
}
\details{
Generated from example data available at \code{system.file("extdata", "01013500_MOD.csv", package = "HyMETT")}
}
\examples{
str(example_mod)

}
\references{
Johnson, M., D. Blodgett, 2020, NOAA National Water Model Reanalysis Data at RENCI, HydroShare,
accessed September 17, 2020 at\cr https://doi.org/10.4211/hs.89b0952512dd4b378dc5be8d2093310f

Johnson, M., 2021, nwmHistoric: National Water Model Historic Data. R package version 0.0.0.9000,
accessed September 17, 2020 at https://github.com/mikejohnson51/nwmHistoric
}
\keyword{datasets}
