% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_preproc.R
\docType{data}
\name{example_preproc}
\alias{example_preproc}
\title{Example Observations prepocessed}
\format{
A data.frame with the following variables:
\describe{
\item{\code{Date}}{}
\item{\code{value}}{}
\item{\code{year}}{}
\item{\code{month}}{}
\item{\code{day}}{}
\item{\code{decimal_date}}{}
\item{\code{WY}}{Water Year: October 1 - September 30}
\item{\code{CY}}{Climate Year: April 1 - March 30}
\item{\code{Q3}}{3-Day Moving Average: computed at end of moving interval}
\item{\code{Q7}}{7-Day Moving Average: computed at end of moving interval}
\item{\code{Q30}}{30-Day Moving Average: computed at end of moving interval}
\item{\code{jd}}{Julian date}
}
}
\usage{
example_preproc
}
\description{
An example dataset with daily observed streamflow preprocessed to include additional timing and
n-day moving averages.
}
\details{
Generated with \code{\link{example_obs}} from \preformatted{
HyMETT::preproc_main(data = example_obs, 
                     Date = "Date", value = "streamflow_cfs", longitude = -68)$daily`}
}
\examples{
str(example_preproc)

}
\seealso{
\code{\link{example_obs}}, \code{\link{preproc_main}}
}
\keyword{datasets}
