% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyMETT-package.R
\name{HyMETT-package}
\alias{HyMETT-package}
\alias{HyMETT}
\alias{_PACKAGE}
\title{Hydrologic Model Evaluation and Time-series Tools}
\description{
Facilitates the analysis and evaluation of hydrologic model output and time-series data with
functions focused on comparison of modeled (simulated) and observed data, period-of-record
statistics, and trends.
}
\details{
Please see \doi{10.5066/P9FNXEWI} for more details.
}
\keyword{package}
